Moved vocab page logic into its own file.
